Reworking rir_types and its list
Due to a bug where a rir type appeared multiple times in the list and
had more subtypes than it should have I had to do quite a bit of rework
of how the rir types list works. Especially the freeing.

Also added a test for the bug that initiated this whole affair.

The downside is that currently rir_types and their list is very
complicated and the logic is really convoluted. This whole area of the
code needs rethinking.