All backend comparisons are always for signed ints
There is no logic to differentiate between signed/unsigned comparison
yet in the RIR or in the backend. LLVM had unsigned comparison by
default which was quite bad.

Turned all comparisons to signed ones.

TODO: Figure out a way to move the choice into the RIR and have it
simply read from the LLVM backend code.